+++ demo/cypress/index.ts
@@ -14,3 +14,4 @@ export { VStepper } from './VStepper';
 export { Wizard } from './Wizard';
 export { DropDown } from './DropDown';
 export { RadioGroup } from './RadioGroup';
+export { MultiSelect } from './MultiSelect';