+++ demo/cypress/MultiSelect/index.tsx
@@ -18,7 +18,17 @@ export const MultiSelect = (): React.ReactElement => {
     'Moscow',
   ];
 
+  const dataObject = [
+    {label: 'London'},
+    {label: 'Islamabad'},
+    {label: 'Berlin'},
+    {label: 'Washington'},
+    {label: 'Paris'},
+  ]
+
   const [value, setValue] = React.useState<string[]>(['London', 'Paris']);
+  const [value1, setValue1] = React.useState<{label: string}[]>();
+  const [value2, setValue2] = React.useState<string[]>();
 
   return(
     <L.Div _demoStory>
@@ -55,6 +65,40 @@ export const MultiSelect = (): React.ReactElement => {
           )
         }}
         value={value}
+      >
+      </L.MultiSelect>
+      <br />
+      <br />
+      <L.P>MultiSelect default dataObject</L.P>
+      <L.MultiSelect
+        id="MSDefaultWithDataObject"
+        compareObjectsBy="label"
+        data={dataObject}
+        textField="label"
+        defaultValue={[{label: 'London'}]}
+        _width40
+        canSelectAll
+        selectAllItemRender={() => 'Select all'}
+        onChange={ev => {
+          setValue1(ev.component.value as {label: string}[]);
+        }}
+        value={value1}
+      >
+      </L.MultiSelect>
+      <br />
+      <br />
+      <L.P>MultiSelect default dataArray</L.P>
+      <L.MultiSelect
+        id="MSDefaultWithDataArray"
+        data={dataArray}
+        textField="label"
+        defaultValue={['London']}
+        _width40
+        canSelectAll
+        onChange={ev => {
+          setValue2(ev.component.value as string[]);
+        }}
+        value={value2}
       >
       </L.MultiSelect>
     </L.Div>