+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.32.0] - 2020-10-05
 
+### Fixed
+- Switcher: Changed wrapper span tag to div
 
 ## [0.31.1] - 2020-09-28